Formatting and Functionality update
Create special debug log file so as not to clutter the regular log when in --debug mode.

--verbose switch added as we are hiding a lot of information

clean command added to clear the local archives

Fixed a few bugs. The real big chagnes are just to fomatting output. We now use a forked pty for dpkg so we can have better control over it's output.