fix(show): not showing `mirror.txt` entries properly
This works for now. I'm going to consider running the filter uri's here and not showing `mirror://` at all and get the real origin's.

Also I would like to consider removing the regex in `filter_uris` with the slice we're using here.

The space not being there is intentional this time as well. Due to the refactor we don't need it