refactor: use `NalaPackage` and `PackageHandler` objects instead of complicated lists
This change is mostly for those who will be looking at or working with the code. Things are much easier to visualize using these objects instead of complicated lists of lists and passing them all around separately.

There may be more refactoring due to this at a later time, but for now this is sufficient. We will be able to make the code very clean using this.

Maybe The main Nala object goes away and we just just use functions. But I'm unsure if I like this idea, although it would be able to clean up __main__ very well.