refactor: move some very used functions into __init__
This commit is a little unorganized. We move some things into init such as the color function.

I cleaned up the term object again. We now disable color if piped, this also solves several bugs related to piping.

There is a perf change here as well, we now check the hash as the file is downloading.

This should also resolve several issues related to piping