fix: not handling broken install during upgrade
fix: held packages showing the wrong things.
refactor: removed redundant check for upgrade mark
refactor: remove dev print

A package could be marked as upgradable, but then it could be marked as removed for the upgrade. In this case Nala would show that there are held packages, but not actually print anything