fix: run apt hooks before handing off to apt
This fix will solve a lot of issues where it's possible for Nala to hang. Essentially we run apt's hooks before we call `DoInstall`, and then we don't allow `apt` to run the hooks.

This commit also changes default to no scrolling windows. It is possible for a program to require user input but the information it has given scrolls out of view.